# Validate frame attributes when setting them, not when getting them