@pep8speaks 2022-07-22
Hello @maxnoe :wave:! It looks like you've made some changes in your pull request, so I've checked the code again for style.


* In the file [`astropy/coordinates/attributes.py`](https://github.com/astropy/astropy/blob/bb563c14659df3854ddfd7a80f62a75568bcb436/astropy/coordinates/attributes.py):

> [Line 136:1](https://github.com/astropy/astropy/blob/bb563c14659df3854ddfd7a80f62a75568bcb436/astropy/coordinates/attributes.py#L136): [W293](https://duckduckgo.com/?q=pep8%20W293) blank line contains whitespace