@maxnoe 2022-07-25
@mhvk Want to take a look? I am also not sure if this is something that can go into 5.2 as @pllim set for now, this might need to wait for 6.0?