@pllim 2022-08-02
Re: 5.2 or 6.0 -- Up to you. If you want it in 6.0, this can never be merged before 5.2.x branching, so safer to turn it back into draft.

Re: breaking sunpy -- Maybe @Cadair can ping his Sunpy peeps to have a look.