@Cadair 2022-08-03
This is definitely going to need some sunpy changes, what worries me from that perspective is it looks pretty hard to support both the old and new code at the same time, which is what sunpy would need to do.

I would need to experiment to see how we could adapt.

@ayshih ping.