@mhvk 2022-08-03
Note that we could avoid repeated validation also by memoizing the outcome. Though I find the logic that is being introduced here, or doing validation on setting, and broadcasting later, a bit more pleasant conceptually. Anyway, it would be good to know in a bit more detail what would go wrong with this in sunpy; we should avoid adding a lot of work if we can solve the performance issue differently.