@Cadair 2022-08-03
Without looking in detail, it's just that we have our Attrs which do validation on get. We would need to adapt and support old and new at the same time. That's as far as I have looked.