@ayshih 2022-08-03
I haven't actually run against this PR, but I suspect we don't need to make any changes in SunPy.  It'll just end up that one of our frame attributes with an overridden `__get__()` will still sometimes validate on getting in addition to on setting (because it can't fully convert the value without accessing a different frame attribute), and I don't see a way around that.