@ayshih 2022-08-04
Now tested: I don't encounter any conflicts with SunPy's existing code.

The changes in this PR are cleaner logic, but I'll point out that the performance improvements are very small.  That's because, in the existing implementation, once the value has been converted and broadcasted on the first call to `__get__()`, it shouldn't be converted or broadcasted again (assuming reasonable definitions of `convert_input()` that bail out quickly).