@maxnoe 2022-09-19
>  That's because, in the existing implementation, once the value has been converted and broadcasted on the first call to __get__(), it shouldn't be converted or broadcasted again (assuming reasonable definitions of convert_input() that bail out quickly).

That's obviously not the case for the ones I tested, see the benchmark code