@mhvk 2022-09-19
For the secondary attribute, I'm not sure we have a test...  Right now, it seems we only use it in `FK4`, for
```
    obstime = TimeAttribute(default=None, secondary_attribute='equinox')
```
validation for `obstime` and `equinox` is the same anyway.

But given that we use it so little, perhaps it is OK to assume validation of the secondary attribute is fine.