@maxnoe 2023-03-20
Yes, I still think this is a good idea, however I cannot promise when I can get back to it